requirements.txt
albumentations==0.4.3
numpy==1.18.1
tqdm==4.41.1
torch==1.4.0
pandas==0.25.1
addict==2.2.1
segmentation_models_pytorch==0.1.0
pytorch_lightning==1.0.6
PyYAML==5.3.1
